+++ Gems/EMotionFX/Code/EMotionFX/CommandSystem/Source/SelectionCommands.cpp
@@ -36,11 +36,11 @@ namespace CommandSystem
     void SelectActorInstancesUsingCommands(const AZStd::vector<EMotionFX::ActorInstance*>& selectedActorInstances)
     {
         SelectionList& selection = GetCommandManager()->GetCurrentSelection();
-        const uint32 numSelectedActorInstances = selectedActorInstances.size();
+        const size_t numSelectedActorInstances = selectedActorInstances.size();
 
         // check if the current selection is equal to the desired actor instances selection list
         bool nothingChanged = true;
-        for (uint32 i = 0; i < numSelectedActorInstances; ++i)
+        for (size_t i = 0; i < numSelectedActorInstances; ++i)
         {
             EMotionFX::ActorInstance* actorInstance = selectedActorInstances[i];
             if (selection.CheckIfHasActorInstance(actorInstance) == false)
@@ -49,7 +49,7 @@ namespace CommandSystem
                 break;
             }
         }
-        for (uint32 i = 0; i < selection.GetNumSelectedActorInstances(); ++i)
+        for (size_t i = 0; i < selection.GetNumSelectedActorInstances(); ++i)
         {
             EMotionFX::ActorInstance* actorInstance = selection.GetActorInstance(i);
             if (AZStd::find(begin(selectedActorInstances), end(selectedActorInstances), actorInstance) == end(selectedActorInstances))
@@ -70,7 +70,7 @@ namespace CommandSystem
 
             // add the newly selected actor instances
             AZStd::string commandString;
-            for (uint32 a = 0; a < numSelectedActorInstances; ++a)
+            for (size_t a = 0; a < numSelectedActorInstances; ++a)
             {
                 EMotionFX::ActorInstance* actorInstance = selectedActorInstances[a];
                 commandString = AZStd::string::format("Select -actorInstanceID %i -actorID %i", actorInstance->GetID(), actorInstance->GetActor()->GetID());
@@ -166,10 +166,10 @@ namespace CommandSystem
         //  return false;
 
         SelectionList& selection        = GetCommandManager()->GetCurrentSelection();
-        const uint32 numActors          = EMotionFX::GetActorManager().GetNumActors();
-        const uint32 numActorInstances  = EMotionFX::GetActorManager().GetNumActorInstances();
-        const uint32 numMotions         = EMotionFX::GetMotionManager().GetNumMotions();
-        const uint32 numAnimGraphs     = EMotionFX::GetAnimGraphManager().GetNumAnimGraphs();
+        const size_t numActors          = EMotionFX::GetActorManager().GetNumActors();
+        const size_t numActorInstances  = EMotionFX::GetActorManager().GetNumActorInstances();
+        const size_t numMotions         = EMotionFX::GetMotionManager().GetNumMotions();
+        const size_t numAnimGraphs     = EMotionFX::GetAnimGraphManager().GetNumAnimGraphs();
 
         AZStd::string valueString;
 
@@ -180,7 +180,7 @@ namespace CommandSystem
             if (AzFramework::StringFunc::Equal(valueString.c_str(), "SELECT_ALL", false /* no case */))
             {
                 // iterate through all available actors and add them to the selection
-                for (uint32 i = 0; i < numActors; ++i)
+                for (size_t i = 0; i < numActors; ++i)
                 {
                     EMotionFX::Actor* actor = EMotionFX::GetActorManager().GetActor(i);
 
@@ -240,7 +240,7 @@ namespace CommandSystem
             }
 
             // iterate through all available actors and add them to the selection
-            for (uint32 i = 0; i < numActors; ++i)
+            for (size_t i = 0; i < numActors; ++i)
             {
                 EMotionFX::Actor* actor = EMotionFX::GetActorManager().GetActor(i);
 
@@ -271,7 +271,7 @@ namespace CommandSystem
             if (AzFramework::StringFunc::Equal(valueString.c_str(), "SELECT_ALL", false /* no case */))
             {
                 // iterate through all available actor instances and add them to the selection
-                for (uint32 i = 0; i < numActorInstances; ++i)
+                for (size_t i = 0; i < numActorInstances; ++i)
                 {
                     EMotionFX::ActorInstance* actorInstance = EMotionFX::GetActorManager().GetActorInstance(i);
 
@@ -330,7 +330,7 @@ namespace CommandSystem
             }
 
             // iterate through all available motions and add them to the selection
-            for (uint32 i = 0; i < numMotions; ++i)
+            for (size_t i = 0; i < numMotions; ++i)
             {
                 // get the current motion
                 EMotionFX::Motion* motion = EMotionFX::GetMotionManager().GetMotion(i);
@@ -362,7 +362,7 @@ namespace CommandSystem
             if (AzFramework::StringFunc::Equal(valueString.c_str(), "SELECT_ALL", false /* no case */))
             {
                 // iterate through all available motions and add them to the selection
-                for (uint32 i = 0; i < numMotions; ++i)
+                for (size_t i = 0; i < numMotions; ++i)
                 {
                     // get the current motion
                     EMotionFX::Motion* motion = EMotionFX::GetMotionManager().GetMotion(i);
@@ -385,7 +385,7 @@ namespace CommandSystem
             else
             {
                 // get the motion index from the string and check if it is valid
-                const uint32 motionIndex = parameters.GetValueAsInt("motionIndex", command);
+                const size_t motionIndex = parameters.GetValueAsInt("motionIndex", command);
                 if (motionIndex >= numMotions)
                 {
                     if (numMotions == 0)
@@ -394,7 +394,7 @@ namespace CommandSystem
                     }
                     else
                     {
-                        outResult = AZStd::string::format("Motion index '%i' is not valid. Valid range is [0, %i].", motionIndex, numMotions - 1);
+                        outResult = AZStd::string::format("Motion index '%zu' is not valid. Valid range is [0, %zu].", motionIndex, numMotions - 1);
                     }
 
                     return false;
@@ -427,7 +427,7 @@ namespace CommandSystem
             if (AzFramework::StringFunc::Equal(valueString.c_str(), "SELECT_ALL", false /* no case */))
             {
                 // iterate through all available motions and add them to the selection
-                for (uint32 i = 0; i < numAnimGraphs; ++i)
+                for (size_t i = 0; i < numAnimGraphs; ++i)
                 {
                     // get the current anim graph
                     EMotionFX::AnimGraph* animGraph = EMotionFX::GetAnimGraphManager().GetAnimGraph(i);
@@ -450,7 +450,7 @@ namespace CommandSystem
             else
             {
                 // get the anim graph index from the string and check if it is valid
-                const uint32 animGraphIndex = parameters.GetValueAsInt("animGraphIndex", command);
+                const size_t animGraphIndex = parameters.GetValueAsInt("animGraphIndex", command);
                 if (animGraphIndex >= numAnimGraphs)
                 {
                     if (numAnimGraphs == 0)
@@ -459,7 +459,7 @@ namespace CommandSystem
                     }
                     else
                     {
-                        outResult = AZStd::string::format("Anim graph index '%i' is not valid. Valid range is [0, %i].", animGraphIndex, numAnimGraphs - 1);
+                        outResult = AZStd::string::format("Anim graph index '%zu' is not valid. Valid range is [0, %zu].", animGraphIndex, numAnimGraphs - 1);
                     }
 
                     return false;
@@ -492,7 +492,7 @@ namespace CommandSystem
             if (AzFramework::StringFunc::Equal(valueString.c_str(), "SELECT_ALL", false /* no case */))
             {
                 // iterate through all available motions and add them to the selection
-                for (uint32 i = 0; i < numAnimGraphs; ++i)
+                for (size_t i = 0; i < numAnimGraphs; ++i)
                 {
                     // get the current anim graph
                     EMotionFX::AnimGraph* animGraph = EMotionFX::GetAnimGraphManager().GetAnimGraph(i);
