+++ Gems/EMotionFX/Code/EMotionFX/Source/AnimGraphObject.h
@@ -134,7 +134,7 @@ namespace EMotionFX
         void InitInternalAttributesForAllInstances();   // does the init for all anim graph instances in the parent animgraph
         virtual void InitInternalAttributes(AnimGraphInstance* animGraphInstance);
         virtual void RemoveInternalAttributesForAllInstances();
-        virtual void DecreaseInternalAttributeIndices(uint32 decreaseEverythingHigherThan);
+        virtual void DecreaseInternalAttributeIndices(size_t decreaseEverythingHigherThan);
 
         virtual void Update(AnimGraphInstance* animGraphInstance, float timePassedInSeconds);
 
@@ -144,14 +144,14 @@ namespace EMotionFX
         virtual void RecursiveOnChangeMotionSet(AnimGraphInstance* animGraphInstance, MotionSet* newMotionSet)   { MCORE_UNUSED(animGraphInstance); MCORE_UNUSED(newMotionSet); }
         virtual void OnActorMotionExtractionNodeChanged()                                                        {}
 
-        MCORE_INLINE uint32 GetObjectIndex() const                                      { return mObjectIndex; }
-        MCORE_INLINE void SetObjectIndex(size_t index)                                  { mObjectIndex = static_cast<uint32>(index); }
+        MCORE_INLINE size_t GetObjectIndex() const                                      { return mObjectIndex; }
+        MCORE_INLINE void SetObjectIndex(size_t index)                                  { mObjectIndex = index; }
 
         MCORE_INLINE AnimGraph* GetAnimGraph() const                                    { return mAnimGraph; }
         MCORE_INLINE void SetAnimGraph(AnimGraph* animGraph)                            { mAnimGraph = animGraph; }
 
-        uint32 SaveUniqueData(AnimGraphInstance* animGraphInstance, uint8* outputBuffer) const;  // save and return number of bytes written, when outputBuffer is nullptr only return num bytes it would write
-        uint32 LoadUniqueData(AnimGraphInstance* animGraphInstance, const uint8* dataBuffer);    // load and return number of bytes read, when dataBuffer is nullptr, 0 should be returned
+        size_t SaveUniqueData(AnimGraphInstance* animGraphInstance, uint8* outputBuffer) const;  // save and return number of bytes written, when outputBuffer is nullptr only return num bytes it would write
+        size_t LoadUniqueData(AnimGraphInstance* animGraphInstance, const uint8* dataBuffer);    // load and return number of bytes read, when dataBuffer is nullptr, 0 should be returned
 
         virtual void RecursiveCollectObjects(AZStd::vector<AnimGraphObject*>& outObjects) const;
 
@@ -167,7 +167,7 @@ namespace EMotionFX
 
     protected:
         AnimGraph*                          mAnimGraph;
-        uint32                              mObjectIndex;
+        size_t                              mObjectIndex;
     };
 } // namespace EMotionFX
